+++ .idea/misc.xml
@@ -11,8 +11,10 @@
         <entry key="..\:/Users/minoro75/AndroidStudioProjects/HereMapsWeatherApp/app/src/main/res/layout/activity_main.xml" value="0.2957427536231884" />
         <entry key="..\:/Users/minoro75/AndroidStudioProjects/HereMapsWeatherApp/app/src/main/res/layout/city_selection_fragment.xml" value="0.28940217391304346" />
         <entry key="..\:/Users/minoro75/AndroidStudioProjects/HereMapsWeatherApp/app/src/main/res/layout/city_weather_fragment.xml" value="0.2957427536231884" />
+        <entry key="..\:/Users/minoro75/AndroidStudioProjects/HereMapsWeatherApp/app/src/main/res/layout/fragment_city_forecast.xml" value="0.20923913043478262" />
         <entry key="..\:/Users/minoro75/AndroidStudioProjects/HereMapsWeatherApp/app/src/main/res/layout/fragment_city_selection.xml" value="0.25" />
         <entry key="..\:/Users/minoro75/AndroidStudioProjects/HereMapsWeatherApp/app/src/main/res/layout/fragment_city_weather.xml" value="0.33" />
+        <entry key="..\:/Users/minoro75/AndroidStudioProjects/HereMapsWeatherApp/app/src/main/res/layout/item_forecast.xml" value="0.21739130434782608" />
         <entry key="..\:/Users/minoro75/AndroidStudioProjects/HereMapsWeatherApp/app/src/main/res/layout/item_weather.xml" value="0.5" />
         <entry key="..\:/Users/minoro75/AndroidStudioProjects/HereMapsWeatherApp/app/src/main/res/layout/main_fragment.xml" value="0.18854166666666666" />
       </map>
