+++ app/src/main/res/layout/fragment_city_forecast.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/surface"
-    tools:context=".ui.city_weather.CityWeatherFragment">
+    tools:context=".ui.city_forecast.CityForecastFragment">
 
     <com.google.android.material.card.MaterialCardView
         android:id="@+id/cardView"
@@ -133,7 +133,7 @@
             android:layout_height="0dp"
             android:background="@color/surface"
 
-            tools:listitem="@layout/item_weather" />
+            tools:listitem="@layout/item_forecast" />
     </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
