+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -27,9 +27,9 @@
 
     <fragment
         android:id="@+id/nav_city_weather"
-        android:name="io.minoro75.demoweatherapp.ui.city_weather.CityWeatherFragment"
+        android:name="io.minoro75.demoweatherapp.ui.city_forecast.CityForecastFragment"
         android:label="City Weather"
-        tools:layout="@layout/fragment_city_weather" >
+        tools:layout="@layout/fragment_city_forecast">
         <argument
             android:name="city"
             app:argType="string" />
